# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Chart API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AlertCondition =
  (typeof AlertCondition)[keyof typeof AlertCondition];

export const AlertCondition = {
  above: "above",
  below: "below",
} as const;

export interface Alert {
  id: string;
  symbol: string;
  price: number;
  condition: AlertCondition;
  active: boolean;
  createdAt: string;
}

export type CreateAlertRequestCondition =
  (typeof CreateAlertRequestCondition)[keyof typeof CreateAlertRequestCondition];

export const CreateAlertRequestCondition = {
  above: "above",
  below: "below",
} as const;

export interface CreateAlertRequest {
  symbol: string;
  price: number;
  condition: CreateAlertRequestCondition;
}

export interface DeleteResponse {
  success: boolean;
}