# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Chart API
 * OpenAPI spec version: 0.1.0
 */
import type { AlertCondition } from "./alertCondition";

export interface Alert {
  id: string;
  symbol: string;
  price: number;
  condition: AlertCondition;
  active: boolean;
  createdAt: string;
}