# lib/api-zod/src/generated/types/createAlertRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Deriv Trading Chart API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAlertRequestCondition } from "./createAlertRequestCondition";

export interface CreateAlertRequest {
  symbol: string;
  price: number;
  condition: CreateAlertRequestCondition;
}